+++ lib/Crypto/Hash/TupleHash256.py
@@ -67,7 +67,4 @@ def new(**kwargs):
 
     custom = kwargs.pop("custom", b'')
 
-    if kwargs:
-        raise TypeError("Unknown parameters: " + str(kwargs))
-
     return TupleHash(custom, cSHAKE256, digest_bytes)