+++ lib/Crypto/Hash/__init__.py
@@ -21,4 +21,5 @@
 __all__ = ['HMAC', 'MD2', 'MD4', 'MD5', 'RIPEMD160', 'SHA1',
            'SHA224', 'SHA256', 'SHA384', 'SHA512', 'CMAC', 'Poly1305',
            'cSHAKE128', 'cSHAKE256', 'KMAC128', 'KMAC256',
-           'TupleHash128', 'TupleHash256', 'KangarooTwelve']
+           'TupleHash128', 'TupleHash256', 'KangarooTwelve',
+           'TurboSHAKE128', 'TurboSHAKE256']