+++ lib/Crypto/SelfTest/Hash/__init__.py
@@ -51,6 +51,7 @@ def get_tests(config={}):
     from Crypto.SelfTest.Hash import test_KMAC;       tests += test_KMAC.get_tests(config=config)
     from Crypto.SelfTest.Hash import test_TupleHash;  tests += test_TupleHash.get_tests(config=config)
     from Crypto.SelfTest.Hash import test_KangarooTwelve;  tests += test_KangarooTwelve.get_tests(config=config)
+    from Crypto.SelfTest.Hash import test_TurboSHAKE; tests += test_TurboSHAKE.get_tests(config=config)
     return tests
 
 if __name__ == '__main__':